requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
dj-database-url==2.0.0
Django==4.2.1
django-bootstrap-v5==1.0.11
django-bootstrap5==23.1
django-heroku==0.3.1
django-mjml==1.1
django-redis==5.2.0
django-taggit==4.0.0
django-user-agents==0.4.0
docopt==0.6.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
msgpack-python==0.5.6
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
PyYAML==6.0.1
redis==4.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==5.5.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
ua-parser==0.18.0
urllib3==2.0.4
user-agents==2.2.0
whitenoise==6.4.0
zope.interface==6.0